refactor(frontend): split the estimate editor into a hook and sections
The first split was not enough: Aikido still read CreateEstimateDialog as
one component doing form state, validation, line editing, totals and the
whole render.

State and the payload move to useEstimateDraft. The form becomes four
sections in EstimateDraftFields - the companion and expiry fields, the
line editor, the notes field and the totals panel. What is left is 103
lines of composition, and each piece can be read on its own.

199 tests still pass, every file in the feature is at 100% statements and
lines, and all 34 Storybook stories still render clean in both themes.
